# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/search": {
      "get": {
        "description": "Perform a comprehensive search across the platform. Supports filtering by type, timeframe, and sorting options.\n    \n    Search Types:\n    - **users**: Search for users by username, name, or bio\n    - **posts**: Search for posts by content or hashtags\n    - **hashtags**: Search for hashtags\n    - **combined**: Search all types (default)\n    \n    Minimum query length: 2 characters",
        "operationId": "SearchController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search query (minimum 2 characters)",
            "schema": {
              "example": "typescript",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Type of content to search (default: combined)",
            "schema": {
              "enum": [
                "users",
                "posts",
                "hashtags",
                "combined"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results per category (default: 20, max: 50)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip (for pagination)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "timeframe",
            "required": false,
            "in": "query",
            "description": "Time window for results (default: all)",
            "schema": {
              "enum": [
                "day",
                "week",
                "month",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sorting method (default: relevance)",
            "schema": {
              "enum": [
                "relevance",
                "popularity",
                "recent"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponseDto"
                      },
                      "description": "Matched users"
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostResponseDto"
                      },
                      "description": "Matched posts"
                    },
                    "hashtags": {
                      "type": "array",
                      "description": "Matched hashtags"
                    },
                    "total": {
                      "type": "number",
                      "description": "Total number of results across all categories",
                      "example": 45
                    },
                    "filters": {
                      "type": "object",
                      "description": "Applied search filters",
                      "properties": {
                        "query": {
                          "type": "string",
                          "example": "typescript"
                        },
                        "type": {
                          "type": "string",
                          "example": "combined"
                        },
                        "limit": {
                          "type": "number",
                          "example": 20
                        },
                        "offset": {
                          "type": "number",
                          "example": 0
                        },
                        "timeframe": {
                          "type": "string",
                          "example": "all"
                        },
                        "sortBy": {
                          "type": "string",
                          "example": "relevance"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Query too short (minimum 2 characters required)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "example": []
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "example": []
                    },
                    "hashtags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "example": []
                    },
                    "total": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "Query must be at least 2 characters long"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Unified search across users, posts, and hashtags",
        "tags": [
          "Search"
        ]
      }
    },
    "/search/suggestions": {
      "get": {
        "description": "Get real-time search suggestions as users type. Returns matching users, hashtags, and popular queries.",
        "operationId": "SearchController_getSuggestions",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search query for suggestions (minimum 2 characters)",
            "schema": {
              "example": "typesc",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of suggestions per category (default: 10, max: 20)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search suggestions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponseDto"
                      },
                      "description": "Suggested users"
                    },
                    "hashtags": {
                      "type": "array",
                      "description": "Suggested hashtags"
                    },
                    "popularQueries": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Popular search queries related to the input",
                      "example": [
                        "#typescript",
                        "@typescript",
                        "typescript tips",
                        "typescript tutorial"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Query too short (minimum 2 characters required)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "example": []
                    },
                    "hashtags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "example": []
                    },
                    "popularQueries": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": []
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get search suggestions",
        "tags": [
          "Search"
        ]
      }
    },
    "/search/trending": {
      "get": {
        "description": "Get currently trending search queries across the platform",
        "operationId": "SearchController_getTrendingSearches",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of trending searches to return (default: 10, max: 50)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trending searches retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trendingSearches": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of trending search queries",
                      "example": [
                        "#javascript",
                        "#react",
                        "@typescript",
                        "web development",
                        "AI"
                      ]
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Last update timestamp",
                      "example": "2024-01-01T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get trending searches",
        "tags": [
          "Search"
        ]
      }
    },
    "/search/discover/trending-posts": {
      "get": {
        "description": "Discover trending posts based on engagement within a specific timeframe",
        "operationId": "SearchController_getTrendingPosts",
        "parameters": [
          {
            "name": "timeframe",
            "required": false,
            "in": "query",
            "description": "Time period for trending calculation (default: day)",
            "schema": {
              "enum": [
                "day",
                "week"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of posts to return (default: 20, max: 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get trending posts for discovery",
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "info": {
    "title": "Echo Social Media API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.echo-app.com",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe123"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL",
            "example": "https://example.com/avatar.jpg"
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "bio": {
            "type": "string",
            "description": "Short biography",
            "example": "Software developer"
          },
          "website": {
            "type": "string",
            "description": "Website URL",
            "example": "https://john.doe.com"
          },
          "location": {
            "type": "string",
            "description": "Location",
            "example": "New York, USA"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Email verification status",
            "example": true
          },
          "resourceType": {
            "type": "string",
            "description": "Resource type",
            "example": "user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "followersCount": {
            "type": "number",
            "description": "Number of followers",
            "example": 150
          },
          "followingCount": {
            "type": "number",
            "description": "Number of followed users",
            "example": 75
          },
          "postCount": {
            "type": "number",
            "description": "Number of posts",
            "example": 42
          },
          "isFollowing": {
            "type": "boolean",
            "description": "Whether current user is following this user",
            "example": false
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "avatar",
          "firstName",
          "lastName",
          "emailVerified",
          "resourceType",
          "createdAt",
          "updatedAt",
          "followersCount",
          "followingCount",
          "postCount",
          "isFollowing"
        ]
      },
...
    }
  }
}